feat(web): extract integrations into a dedicated /integrations page
Move Trakt and third-party plugin connections out of the profile page
into a dedicated IntegrationsLive at /integrations, with a nav entry and
its own test. Profile keeps account and preferences only.
